+++ test/test_tasks.cjs
@@ -1,5 +1,5 @@
 const { expect } = require("chai");
-const { describe, it: test } = require("mocha");
+const { it } = require("mocha");
 const childProcess = require("child_process");
 
 // helper method to run command in shell
@@ -11,8 +11,7 @@ const run = (args) => {
     return exec(`npx udc-js ${args}`);
 };
 
-test("should talk about Tasks", () => {
+it("should talk about Tasks", () => {
     const output = run(`list`);
     expect(output).to.contain("Tasks");
 });
-
